Make a general base class for all elements, incl desing elements
