+++ marxs/optics/base.py
@@ -1,53 +1,12 @@
 # import abc  - do I need this?
-from collections import OrderedDict
 from functools import wraps
 from copy import copy
 
 import numpy as np
-from transforms3d import affines
 from astropy.table import Table, Row, Column
 
 from ..math.pluecker import *
-
-
-class SimulationSequenceElement(object):
-    '''Base element for everything that's allowed in a sequence. Could go into
-    a more general file once it gets more than 5 lines as it's also used for
-    e.g. source and maybe also filters, files handles?
-    '''
-    def __init__(self, **kwargs):
-        if 'name' in kwargs:
-            self.name = kwargs.pop('name')
-        else:
-            self.name = self.__class__
-
-        if len(kwargs) > 0:
-            raise ValueError('Initialization arguments {0} not understood'.format(', '.join(kwargs.keys())))
-
-    def describe(self):
-        return OrderedDict(element=self.name)
-
-
-def _parse_position_keywords(kwargs):
-    '''Parse keywords to define position.
-
-    If pos4d is given, use that, otherwise look for ``position``, ``zoom``
-    and ``orientation``.
-    '''
-    pos4d = kwargs.pop('pos4d', None)
-    if pos4d is None:
-        position = kwargs.pop('position', np.zeros(3))
-        orientation = kwargs.pop('orientation', np.eye(3))
-        zoom = kwargs.pop('zoom', 1.)
-        if np.isscalar(zoom):
-            zoom = np.ones(3) * zoom
-        if not len(zoom) == 3:
-            raise ValueError('zoom must have three elements for x,y,z or be a scalar (global zoom).')
-        pos4d = affines.compose(position, orientation, zoom)
-    else:
-        if ('position' in kwargs) or ('orientation' in kwargs) or ('zoom' in kwargs):
-            raise ValueError('If pos4d is specificed, the following keywords cannot be given at the same time: position, orientation, zoom.')
-    return pos4d
+from ..base import SimulationSequenceElement, _parse_position_keywords
 
 class OpticalElement(SimulationSequenceElement):
     '''
